refactor: move ConversationTrigger to app_conversation_models and remove unused data models
- Move ConversationTrigger class from conversation_metadata.py to app_conversation_models.py
- Re-export ConversationTrigger from conversation_metadata.py for backward compatibility
- Delete unused conversation_metadata_result_set.py
- Delete unused conversation_status.py
