+++ openhands/storage/data_models/conversation_metadata.py
@@ -1,22 +1,13 @@
 from dataclasses import dataclass, field
 from datetime import datetime, timezone
-from enum import Enum
 
+from openhands.app_server.app_conversation.app_conversation_models import (
+    ConversationTrigger,
+)
 from openhands.integrations.service_types import ProviderType
 
-
-class ConversationTrigger(Enum):
-    RESOLVER = 'resolver'
-    GUI = 'gui'
-    SUGGESTED_TASK = 'suggested_task'
-    REMOTE_API_KEY = 'openhands_api'
-    SLACK = 'slack'
-    MICROAGENT_MANAGEMENT = 'microagent_management'
-    JIRA = 'jira'
-    JIRA_DC = 'jira_dc'
-    LINEAR = 'linear'
-    BITBUCKET = 'bitbucket'
-    AUTOMATION = 'automation'
+# Re-export for backward compatibility
+__all__ = ['ConversationTrigger', 'ConversationMetadata']
 
 
 @dataclass
