--- openhands/storage/data_models/conversation_metadata_result_set.py
@@ -1,9 +1,0 @@
-from dataclasses import dataclass, field
-
-from openhands.storage.data_models.conversation_metadata import ConversationMetadata
-
-
-@dataclass
-class ConversationMetadataResultSet:
-    results: list[ConversationMetadata] = field(default_factory=list)
-    next_page_id: str | None = None
